Allow more than one ToDo entry
Previously the todo itself was added and this bound the last item on the list to the todo entered on the form.  This meant that if you tried to add another todo, the last todo on the list was edited, but no new todos were created.
controller.js  Add a copy to the todo to the list of todos.
               On successful save of the todo -> reset the form fields.
todo.html      Set the title field as an autofocus field
               Change the option value from "ToDo" to "Todo" which is required by the server that saves the todos.
               Added a script to set the autofocus when the modal is shown.