Allow Editing & Deleting of recent entries
Editing and Deleting rely on passing the id of the ToDo item to the server.
The id is generated by the server when the ToDo item is created.
This id is unknown to the ToDo list for newly created items until the ToDo list is fetched from the server.
controller.js  Added call to getToDos() method of the TodoService after creating a new ToDo item.